requirements.txt
django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9        # Only if using PostgreSQL
python-decouple>=3.8        # Optional: environment variables
django-environ
django-cors-headers>=4.0
gunicorn
djangorestframework-simplejwt
requests
psycopg[binary,pool]
oracledb
opencv-python-headless
numpy
mediapipe
cx_Oracle==8.3
python-dotenv
